+++ app/src/main/java/com/github/ipcjs/screenshottile/OnAcquireScreenshotPermissionListener.java
@@ -1,5 +1,0 @@
-package com.github.ipcjs.screenshottile;
-
-public interface OnAcquireScreenshotPermissionListener {
-    void onAcquireScreenshotPermission(boolean isNewPermission);
-}